fixed timing integer truncation

+++ src/ofxTLElement.cpp
@@ -180,21 +180,21 @@ int ofxTLElement::screenXForIndex(int index, int durationInFrames){
 	return ofMap(index, startFrame, endFrame, bounds.x, bounds.x+bounds.width, false);
 }
 
-int ofxTLElement::screenXForTime(float time){
+float ofxTLElement::screenXForTime(float time){
 	return screenXForTime(time, timeline->getDurationInSeconds());
 }
 
-float ofxTLElement::timeForScreenX(int screenX){
+float ofxTLElement::timeForScreenX(float screenX){
 	return timeForScreenX(screenX, timeline->getDurationInSeconds());
 }
 
-int ofxTLElement::screenXForTime(float time, float durationInSeconds){
+float ofxTLElement::screenXForTime(float time, float durationInSeconds){
 	float startTime = zoomBounds.min * durationInSeconds;
 	float endTime = zoomBounds.max *durationInSeconds;
 	return ofMap(time, startTime, endTime, bounds.x, bounds.x+bounds.width, false);
 }
 
-float ofxTLElement::timeForScreenX(int screenX, float durationInSeconds){
+float ofxTLElement::timeForScreenX(float screenX, float durationInSeconds){
 	float startTime = zoomBounds.min * durationInSeconds;
 	float endTime = zoomBounds.max *durationInSeconds;
 	return ofMap(screenX, bounds.x, bounds.x+bounds.width, startTime, endTime, true);
