+++ src/ofxTLElement.h
@@ -120,10 +120,10 @@ class ofxTLElement
 	virtual int indexForScreenX(int screenX, int durationInFrames);
 	virtual int screenXForIndex(int index, int durationInFrames);
 	
-	virtual int screenXForTime(float time);
-	virtual float timeForScreenX(int screenX);
-	virtual int screenXForTime(float time, float durationInSeconds);
-	virtual float timeForScreenX(int screenX, float durationInSeconds);
+	virtual float screenXForTime(float time);
+	virtual float timeForScreenX(float screenX);
+	virtual float screenXForTime(float time, float durationInSeconds);
+	virtual float timeForScreenX(float screenX, float durationInSeconds);
 		
 	virtual void drawRectChanged(){};
